updated code for git-local branch

+++ controllers/resourceController.js
@@ -1,8 +1,8 @@
-// const { Permit } = require("permitio");
+const { Permit } = require("permitio");
 const { Resource } = require("../models/schema");
 const { User } = require("../models/schema");
 const ExpressError = require('../utils/ExpressError');
-// const permitAuthorization = require('../utils/permitAuthorization');
+const permitAuthorization = require('../utils/permitAuthorization');
 
 function createResource(key, value, user, symmetricKey) {
   return new Promise((resolve, reject) => {
@@ -17,10 +17,10 @@ function createResource(key, value, user, symmetricKey) {
           resourceOwner: user._id,
         });
         // make permit resource
-        // await permitAuthorization.createPermitResource(resource._id, resource.resourceName).then(async ()=>{
-        //   // assign user its role
-        //   await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "owner");
-        // })
+        await permitAuthorization.createPermitResource(resource._id, resource.resourceName).then(async ()=>{
+          // assign user its role
+          await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "owner");
+        })
     
         resolve(userOwnedResources);
       })
@@ -55,11 +55,11 @@ module.exports.createResource = async (req, res) => {
 
 module.exports.deleteResource = async (req, res) => {
   const { resourceId } = req.body;
-  // const check = await permitAuthorization.checkPermission(req.userDetails.username, "delete", resourceId, "cred");
-  // if(!check.check) return res.status(403).json({
-  //     deleted: false,
-  //     msg: "User not permitted"
-  // });
+  const check = await permitAuthorization.checkPermission(req.userDetails.username, "delete", resourceId, "cred");
+  if(!check.check) return res.status(403).json({
+      deleted: false,
+      msg: "User not permitted"
+  });
   Resource.deleteOne({ _id: resourceId , 
     $or:[
       {resourceOwner: req.userDetails._id },
@@ -74,7 +74,7 @@ module.exports.deleteResource = async (req, res) => {
           msg: "User not permitted"
         });
       }
-      // await permitAuthorization.deleteResource(resourceId);
+      await permitAuthorization.deleteResource(resourceId);
       User.updateMany({},{ $pull: {encryptedSymmetricKeys: { resourceId: resourceId }}})
       .then(result => {
         return res.status(200).json({
@@ -115,8 +115,8 @@ module.exports.addMemberToResource = async (req, res) => {
   //   console.log("You are the OWNER");
   //   return res.status(200).json({ response: 2 });
   // }
-  // const permitCheck = await permitAuthorization.checkPermission(user.username, "share", resource._id, "cred");
-  // if(permitCheck.check) return res.status(200).json({ response: 2 });
+  const permitCheck = await permitAuthorization.checkPermission(user.username, "share", resource._id, "cred");
+  if(permitCheck.check) return res.status(200).json({ response: 2 });
 
   // check for role: add, remove others
 
@@ -133,14 +133,14 @@ module.exports.addMemberToResource = async (req, res) => {
       });
       resource.save();
 
-      // if(resource.editors.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "editor");
-      // if(resource.authors.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "author");
+      if(resource.editors.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "editor");
+      if(resource.authors.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "author");
       
       console.log(`user: ${user.username} is now VIEWER of: ${resourceId}/${resource.resourceName} after removing from others`);
-      // await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "viewer").then(()=> {
-      //   return res.status(200).json({ response: 1 });
-      //  })
-      return res.status(200).json({ response: 1 });
+      await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "viewer").then(()=> {
+        return res.status(200).json({ response: 1 });
+       })
+      // return res.status(200).json({ response: 1 });
     } else {
       await resource.updateOne({
         $push: { viewers: user._id },
@@ -148,10 +148,10 @@ module.exports.addMemberToResource = async (req, res) => {
       });
       resource.save();
       console.log(`user: ${user.username} is NOW viewer of: ${resourceId}/${resource.resourceName}`);
-      // await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "viewer").then(()=> {
-      //   return res.status(200).json({ response: 1 });
-      //  })
-      return res.status(200).json({ response: 1 });
+      await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "viewer").then(()=> {
+        return res.status(200).json({ response: 1 });
+       })
+      // return res.status(200).json({ response: 1 });
     }
   } else if (role == "editor") {
     console.log("--------------- A REQUEST IS SENT TO EDITOR ---------------");
@@ -166,14 +166,14 @@ module.exports.addMemberToResource = async (req, res) => {
       });
       resource.save();
 
-      // if(resource.viewers.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "viewer");
-      // if(resource.authors.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "author");
+      if(resource.viewers.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "viewer");
+      if(resource.authors.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "author");
 
       console.log(`user: ${user.username} is now EDITOR of: ${resourceId}/${resource.resourceName} after removing from others`);
-      // await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "editor").then(()=> {
-      //   return res.status(200).json({ response: 1 });
-      //  })
-       return res.status(200).json({ response: 1 });
+      await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "editor").then(()=> {
+        return res.status(200).json({ response: 1 });
+       })
+      //  return res.status(200).json({ response: 1 });
     } else {
       await resource.updateOne({
         $push: { editors: user._id },
@@ -181,10 +181,10 @@ module.exports.addMemberToResource = async (req, res) => {
       });
       resource.save();
       console.log(`user: ${user.username} is NOW editor of: ${resourceId}/${resource.resourceName}`);
-      // await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "editor").then(()=> {
-      //   return res.status(200).json({ response: 1 });
-      //  })
-      return res.status(200).json({ response: 1 });
+      await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "editor").then(()=> {
+        return res.status(200).json({ response: 1 });
+       })
+      // return res.status(200).json({ response: 1 });
     }
   } else if (role == "author") {
     console.log("--------------- A REQUEST IS SENT TO AUTHOR ---------------");
@@ -199,14 +199,14 @@ module.exports.addMemberToResource = async (req, res) => {
       });
       resource.save();
 
-      // if(resource.editors.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "editor");
-      // if(resource.viewers.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "viewer");
+      if(resource.editors.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "editor");
+      if(resource.viewers.includes(user._id)) await permitAuthorization.unassignPermitRole(user.username, resource._id, "viewer");
 
       console.log(`user: ${user.username} is now AUTHOR of: ${resourceId}/${resource.resourceName} after removing from others`);
-      // await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "author").then(()=> {
-      //   return res.status(200).json({ response: 1 });
-      //  })
-      return res.status(200).json({ response: 1 });
+      await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "author").then(()=> {
+        return res.status(200).json({ response: 1 });
+       })
+      // return res.status(200).json({ response: 1 });
     } else {
       await resource.updateOne({
         $push: { authors: user._id },
@@ -214,10 +214,10 @@ module.exports.addMemberToResource = async (req, res) => {
       });
       resource.save();
       console.log(`user: ${user.username} is NOW author of: ${resourceId}/${resource.resourceName}`);
-      // await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "author").then(()=> {
-      //  return res.status(200).json({ response: 1 });
-      // })
-      return res.status(200).json({ response: 1 });
+      await permitAuthorization.createPermitResourceInstanceAndAssignRole(user.username, resource._id, "author").then(()=> {
+       return res.status(200).json({ response: 1 });
+      })
+      // return res.status(200).json({ response: 1 });
     }
   } else {
     console.log("Not a Role T_T");
@@ -232,11 +232,11 @@ module.exports.editResource = async (req, res) => {
       response: 0,
     });
   }
-  // const permitResult = await permitAuthorization.checkPermission(req.userDetails.username, "edit", resId, "cred");
-  // if(!permitResult) return res.status(403).json({ 
-  //   response: 2,
-  //   msg: "Not permitted"
-  // });
+  const permitResult = await permitAuthorization.checkPermission(req.userDetails.username, "edit", resId, "cred");
+  if(!permitResult) return res.status(403).json({ 
+    response: 2,
+    msg: "Not permitted"
+  });
   try{
   Resource.findOneAndUpdate(
     { _id: resId , 
@@ -291,7 +291,7 @@ module.exports.showSharedResources = async (req, res) => {
     // Create an array of promises
     const promises = resourcesSharedWithUser.map(async (resource) => {
       if (resource.viewers.includes(req.userDetails._id)) {
-        // const checkViewer = await permitAuthorization.checkPermission( req.userDetails.username, "view", resource._id, "cred" );
+        const checkViewer = await permitAuthorization.checkPermission( req.userDetails.username, "view", resource._id, "cred" );
         const result = req.userDetails.encryptedSymmetricKeys.find(obj => obj.resourceId.toString() === resource._id.toString() );
         resourcesToSend.viewer.push({
           resourceName: resource.resourceName,
@@ -301,9 +301,9 @@ module.exports.showSharedResources = async (req, res) => {
           _id: resource._id,
           encryptedSymmetricKey: result.encryptedSymmetricKey,
         });
-        // if (!checkViewer.check) resourcesToSend.viewer.length = 0;
+        if (!checkViewer.check) resourcesToSend.viewer.length = 0;
       } else if (resource.editors.includes(req.userDetails._id)) {
-        // const checkEditor = await permitAuthorization.checkPermission( req.userDetails.username, "edit", resource._id, "cred" );
+        const checkEditor = await permitAuthorization.checkPermission( req.userDetails.username, "edit", resource._id, "cred" );
         const result = req.userDetails.encryptedSymmetricKeys.find(obj => obj.resourceId.toString() === resource._id.toString() );
         resourcesToSend.editor.push({
           resourceName: resource.resourceName,
@@ -313,9 +313,9 @@ module.exports.showSharedResources = async (req, res) => {
           _id: resource._id,
           encryptedSymmetricKey: result.encryptedSymmetricKey,
         });
-        // if (!checkEditor.check) resourcesToSend.editor.length = 0;
+        if (!checkEditor.check) resourcesToSend.editor.length = 0;
       } else {
-        // const checkAuthor = await permitAuthorization.checkPermission( req.userDetails.username, "delete", resource._id, "cred" );
+        const checkAuthor = await permitAuthorization.checkPermission( req.userDetails.username, "delete", resource._id, "cred" );
         const result = req.userDetails.encryptedSymmetricKeys.find(obj => obj.resourceId.toString() === resource._id.toString() );
         resourcesToSend.author.push({
           resourceName: resource.resourceName,
@@ -325,7 +325,7 @@ module.exports.showSharedResources = async (req, res) => {
           _id: resource._id,
           encryptedSymmetricKey: result.encryptedSymmetricKey,
         });
-        // if (!checkAuthor.check) resourcesToSend.author.length = 0;
+        if (!checkAuthor.check) resourcesToSend.author.length = 0;
       }
     });
 
@@ -347,8 +347,8 @@ module.exports.showResourceInfo = async (req, res) => {
     if(!resource) return res.json({
       msg:"Resource not found (┬┬﹏┬┬)"
     })
-    // const check = await permitAuthorization.checkPermission(req.userDetails.username, "share", resource._id, "cred");
-    // if(!check.check) return res.status(401).json({ msg: "unauthorized to view the resource info" });
+    const check = await permitAuthorization.checkPermission(req.userDetails.username, "share", resource._id, "cred");
+    if(!check.check) return res.status(401).json({ msg: "unauthorized to view the resource info" });
     // console.log(resource)
     const resourceToSend = { viewers:[], editors:[], authors:[] }
     resource.viewers.forEach(res => {
@@ -370,8 +370,8 @@ module.exports.removeResourcePermission = async (req, res) => {
   User.findOne({username: username}).then(async user => {
     if(!user) return res.json({ msg: "User not found ━┳━ ━┳━", remove: false });
     
-    // const check = await permitAuthorization.checkPermission(req.userDetails.username, "share", resId, "cred");
-    // if(!check.check) return res.status(401).json({ msg: "unauthorized to remove the permissions" });
+    const check = await permitAuthorization.checkPermission(req.userDetails.username, "share", resId, "cred");
+    if(!check.check) return res.status(401).json({ msg: "unauthorized to remove the permissions" });
 
     await user.updateOne({ $pull: { encryptedSymmetricKeys: { resourceId: resId } } });
     user.save();
@@ -388,7 +388,7 @@ module.exports.removeResourcePermission = async (req, res) => {
         resource.resourceSharedWith = false;
         await resource.save();
       }
-      // await permitAuthorization.unassignPermitRole(username, resId, "viewer");
+      await permitAuthorization.unassignPermitRole(username, resId, "viewer");
       return res.status(200).json({
         msg: "Removed permission successfully （￣︶￣）↗　",
         remove: true
@@ -410,7 +410,7 @@ module.exports.removeResourcePermission = async (req, res) => {
         resource.resourceSharedWith = false;
         await resource.save();
       }
-      // await permitAuthorization.unassignPermitRole(username, resId, "author");
+      await permitAuthorization.unassignPermitRole(username, resId, "author");
       return res.status(200).json({
         msg: "Removed permission successfully （￣︶￣）↗　",
         remove: true
@@ -432,7 +432,7 @@ module.exports.removeResourcePermission = async (req, res) => {
         resource.resourceSharedWith = false;
         await resource.save();
       }
-      // await permitAuthorization.unassignPermitRole(username, resId, "editor");
+      await permitAuthorization.unassignPermitRole(username, resId, "editor");
       return res.status(200).json({
         msg: "Removed permission successfully （￣︶￣）↗　",
         remove: true
